Add files via upload
Write a program to print the final scores of two students by creating a class
named ‘FinalScore'. The class takes the values of its coursework and final exam both out of
100 as parameters of its constructor. The class should have a method named 'returnFinal'
which returns the final course score of a student by adding the coursework and final exam
scores using the coursework to final exam ratio of 40 to 60. The values of coursework and
final exam scores for each student are assigned in the main method.